// src/common/enums/blog-message.enum.ts
export enum BlogMessages {
    NOT_FOUND = 'Blog not found',
    ALREADY_EXISTS = 'Blog already exists',
    CREATED = 'Blog created successfully',
    UPDATED = 'Blog updated successfully',
    DELETED = 'Blog deleted successfully',

    UNAUTHORIZED_ACCESS = 'Unauthorized access',
    INVALID_CREDENTIALS = 'Invalid credentials',
}